#ESW-489: Shivani | Manish: Move logic regarding step list & Sequencer state to sequencer details component
note : not using useStream for this use case
