+++ src/features/sequencer/components/SequencersTable.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 import { HeaderTitle } from '../../../components/table/HeaderTitle'
 import { getSequencerPath } from '../../../routes/RoutesConfig'
-import type { SequencerInfo } from '../hooks/useSequencersData'
+import type { SequencerInfo } from '../utils'
 import styles from './sequencerTable.module.css'
 import type { ColumnsType } from 'antd/lib/table/interface'
 import type { BaseType } from 'antd/lib/typography/Base'
