+++ src/features/sequencer/hooks/useSequencerDetails.ts
@@ -3,19 +3,21 @@ import { useStream } from '../../../hooks/useStream'
 import { useSequencerService } from './useSequencerService'
 import type { Prefix, SequencerStateResponse } from '@tmtsoftware/esw-ts'
 
-const onEvent = (sequencerStateResponse: SequencerStateResponse) =>
-  sequencerStateResponse
-
 export const useSequencerDetails = (
   sequencerPrefix: Prefix
 ): [SequencerStateResponse | undefined, boolean] => {
   console.log('using sequencer service:')
   const sequencerService = useSequencerService(sequencerPrefix)
-  if (!sequencerService) throw Error('sequencer not found sorry')
+  if (!sequencerService) throw Error('sequencer not found')
 
   const subscribeState = useCallback(
     (onEvent: (sequencerStateResponse: SequencerStateResponse) => void) =>
       sequencerService.subscribeSequencerState()(onEvent),
+    []
+  )
+
+  const onEvent = useCallback(
+    (sequencerStateResponse: SequencerStateResponse) => sequencerStateResponse,
     []
   )
